# front/src/shared/api/g/api/models/ConvertCurrencyModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * back, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Currency } from './Currency';
import {
    CurrencyFromJSON,
    CurrencyFromJSONTyped,
    CurrencyToJSON,
} from './Currency';

/**
 * 
 * @export
 * @interface ConvertCurrencyModel
 */
export interface ConvertCurrencyModel {
    /**
     * 
     * @type {number}
     * @memberof ConvertCurrencyModel
     */
    accountId?: number;
    /**
     * 
     * @type {Currency}
     * @memberof ConvertCurrencyModel
     */
    targetCurrency?: Currency;
}

/**
 * Check if a given object implements the ConvertCurrencyModel interface.
 */
export function instanceOfConvertCurrencyModel(value: object): value is ConvertCurrencyModel {
    return true;
}

export function ConvertCurrencyModelFromJSON(json: any): ConvertCurrencyModel {
    return ConvertCurrencyModelFromJSONTyped(json, false);
}

export function ConvertCurrencyModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConvertCurrencyModel {
    if (json == null) {
        return json;
    }
    return {
        
        'accountId': json['accountId'] == null ? undefined : json['accountId'],
        'targetCurrency': json['targetCurrency'] == null ? undefined : CurrencyFromJSON(json['targetCurrency']),
    };
}

export function ConvertCurrencyModelToJSON(value?: ConvertCurrencyModel | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'accountId': value['accountId'],
        'targetCurrency': CurrencyToJSON(value['targetCurrency']),
    };
}